+++ src/main/java/org/plumelib/util/MPair.java
@@ -61,9 +61,9 @@ public final class MPair<V1, V2> {
   public static <T1 extends Cloneable, T2 extends Cloneable> MPair<T1, T2> cloneElements(
       MPair<T1, T2> orig) {
     T1 oldFirst = orig.first;
-    T1 newFirst = oldFirst == null ? oldFirst : UtilPlume.clone(oldFirst);
+    T1 newFirst = oldFirst == null ? oldFirst : UtilP.clone(oldFirst);
     T2 oldSecond = orig.second;
-    T2 newSecond = oldSecond == null ? oldSecond : UtilPlume.clone(oldSecond);
+    T2 newSecond = oldSecond == null ? oldSecond : UtilP.clone(oldSecond);
     return of(newFirst, newSecond);
   }