+++ src/test/java/org/plumelib/util/GraphPTest.java
@@ -14,9 +14,9 @@ import org.junit.jupiter.api.Test;
   "keyfor", // https://github.com/typetools/checker-framework/issues/2358 and maybe other issues
   "nullness"
 })
-final class GraphPlumeTest {
+final class GraphPTest {
 
-  GraphPlumeTest() {}
+  GraphPTest() {}
 
   // ///////////////////////////////////////////////////////////////////////////
 
@@ -54,11 +54,11 @@ final class GraphPlumeTest {
   }
 
   @Test
-  void testGraphPlume() {
+  void testGraphP() {
 
     initializePreds1AndSucc1();
 
-    Map<Integer, List<Integer>> dom1post = GraphPlume.dominators(succs1);
+    Map<Integer, List<Integer>> dom1post = GraphP.dominators(succs1);
     assertEquals("[7, 1, 0]", dom1post.get(0).toString());
     assertEquals("[7, 1]", dom1post.get(1).toString());
     assertEquals("[7, 2]", dom1post.get(2).toString());
@@ -68,7 +68,7 @@ final class GraphPlumeTest {
     assertEquals("[7, 5, 4, 6]", dom1post.get(6).toString());
     assertEquals("[7]", dom1post.get(7).toString());
 
-    Map<Integer, List<Integer>> dom1pre = GraphPlume.dominators(preds1);
+    Map<Integer, List<Integer>> dom1pre = GraphP.dominators(preds1);
     assertEquals("[0]", dom1pre.get(0).toString());
     assertEquals("[0, 1]", dom1pre.get(1).toString());
     assertEquals("[0, 1, 2]", dom1pre.get(2).toString());