+++ src/test/java/org/plumelib/util/SystemPTest.java
@@ -1,0 +1,9 @@
+package org.plumelib.util;
+
+final class SystemPTest {
+
+  SystemPTest() {}
+
+  // No tests. :-(
+
+}